правки по комментариям с предыдущего дз
Вынесен один модуль user block, добавлен переход в MyList по аватарке
