+++ project/src/components/add-review/add-review-screen.tsx
@@ -2,13 +2,10 @@ import Logo from '../logo/logo';
 import {ChangeEvent, useState} from 'react';
 import {useSelector} from 'react-redux';
 import {State} from '../../types/state';
-import { AuthorizationStatus } from '../../const';
-import UserBlockLoggedIn from '../user-block/user-block-logged-in';
-import UserBlockLoggedOut from '../user-block/user-block-logged-out';
+import {UserBlock} from '../user-block/user-block';
 
 function AddReviewScreen(): JSX.Element {
   const films = useSelector((state: State) => state.films);
-  const auth = useSelector((state: State) => state.authorizationStatus);
   const [firstFilm] = films;
   const {title, image, poster} = firstFilm;
   const [review, setReview] = useState('');
@@ -39,8 +36,7 @@ function AddReviewScreen(): JSX.Element {
               </li>
             </ul>
           </nav>
-
-          {auth === AuthorizationStatus.Auth ? <UserBlockLoggedIn /> : <UserBlockLoggedOut />}
+          <UserBlock/>
         </header>
 
         <div className="film-card__poster film-card__poster--small">
