+++ project/src/components/footer/footer.tsx
@@ -8,7 +8,7 @@ function Footer(): JSX.Element {
       </div>
 
       <div className="copyright">
-        <p>© 2019 What to watch Ltd.</p>
+        <p>© 2021 What to watch Ltd.</p>
       </div>
     </footer>
   );
