+++ project/src/components/movie-page-details/movie-page-details-screen.tsx
@@ -1,15 +1,13 @@
 import {Link} from 'react-router-dom';
-import {AppRoute, AuthorizationStatus} from '../../const';
+import {AppRoute} from '../../const';
 import Logo from '../logo/logo';
 import {generatePath} from 'react-router';
 import {useSelector} from 'react-redux';
 import {State} from '../../types/state';
-import UserBlockLoggedIn from '../user-block/user-block-logged-in';
-import UserBlockLoggedOut from '../user-block/user-block-logged-out';
+import {UserBlock} from '../user-block/user-block';
 
 function MoviePageDetailsScreen(): JSX.Element {
   const films = useSelector((state: State) => state.films);
-  const auth = useSelector((state: State) => state.authorizationStatus);
 
   const [firstFilm] = films;
   const {poster, image, title, genre, releaseYear, director, runtime, starring} = firstFilm;
@@ -29,8 +27,7 @@ function MoviePageDetailsScreen(): JSX.Element {
               <Logo />
             </div>
 
-            {auth === AuthorizationStatus.Auth ? <UserBlockLoggedIn /> : <UserBlockLoggedOut />}
-
+            <UserBlock/>
           </header>
 
           <div className="film-card__wrap">
